feat(pdd): unified RBLN NIXL connector — host-bounce + D2D over RDMA, DT-only
A single RblnNixlConnector + RblnNixlConnectorWorker handles both
P/D paths, dispatched by kv_transfer_config.kv_buffer_device:

  "cpu"  → host-bounce: page-aligned CPU staging (`_aligned_like`
           matches shape *and* dtype), RDMA over DRAM_SEG via the
           RBLN NIXL backend's ibv_reg_mr path.
  "rbln" → D2D: VRAM_SEG via the nixl_rbln adapter's dmabuf path;
           no host staging; registration deferred until warm-up
           materializes the KV cache backing memory.

Both paths transport over the same RBLN backend / RDMA NICs (no
UCX/TCP fallback) and require VLLM_RBLN_USE_DEVICE_TENSOR=1.

Runner (`rbln_copy_kv_blocks` closure): a single torch path that
splits K/V (dim 0) first so each per-block view is contiguous,
hitting `_copy_from_rbln`'s direct-DMA fast path. No staging, no
`.to(device)` allocation — bytes flow directly between existing
host and device storage. Measured on Qwen3-32B P/D (4 blocks/xfer):
prefill_d2h 30 ms, decode_h2d 31 ms — PCIe Gen4 x16 line rate.

Worker class:
* nixl_backends = ["RBLN"]; D2D additionally sets nixl_memory_type
  = "VRAM".
* _runtime_holder / _pending_kv_caches declared on both paths so
  the connector's set_runtime_holder / finalize_kv_cache_registration
  pass through without getattr probes.
* register_kv_caches branches on kv_buffer_device: D2D stashes for
  defer, host-bounce calls ensure_rbln_backend (device-keyed ctx
  lookup) then super.
* _register_kv_caches_impl carries the D2D nixl_rbln.register_kv_regions
  flow (TpKVTopology, region collection, MR registration, transfer
  table wire-up).

Bookkeeping:
* RblnNixlDirectConnector and its factory registration removed.
* Worker rbln_worker.py comment updated to reference the unified
  connector's D2D path.
