+++ vllm_rbln/v1/worker/rbln_worker.py
@@ -500,9 +500,9 @@ class RBLNWorker(WorkerBase):
 
                 raise
 
-        # The direct-vmem KV connector (RblnNixlDirectConnector) defers its
-        # NIXL registration until the KV cache physical views exist, which
-        # only happens once warm-up has run the compiled model.
+        # The D2D path of RblnNixlConnector (kv_buffer_device="rbln") defers
+        # its NIXL registration until the KV cache physical views exist,
+        # which only happens once warm-up has run the compiled model.
         if has_kv_transfer_group():
             connector = get_kv_transfer_group()
             finalize = getattr(connector, "finalize_kv_cache_registration", None)
